% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center variables on their mean}
\usage{
center(data, ..., na.rm = TRUE)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{...}{variables to center. Can use \code{tidyselect} as in \code{\link[dplyr:select]{dplyr::select()}}.
If none are included, default is to center all numeric variables.}
}
\value{
a \code{data.frame}
}
\description{
Center variables on their mean
}
\examples{

library(dplyr)

center(mtcars)

center(mtcars, hp)

center(mtcars, starts_with("c"))

}
