% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatrs.R
\name{clean_pval}
\alias{clean_pval}
\title{Clean p-value}
\usage{
clean_pval(x, less_than = 0.001, digits = 3, add_stars = FALSE,
  stars = list(pval = c(0.1, 0.05, 0.01, 0.001), symbol = c(".", "*",
  "**", "***")))
}
\arguments{
\item{x}{a vector of p-values}

\item{less_than}{the lower bound for printing complete numbers}

\item{digits}{the number of digits to round to}

\item{add_stars}{logical. Add stars to the p-value?}

\item{stars}{a list of p-values cut-offs and matching symbols}
}
\value{
a character vector
}
\description{
This function cleans a p-value by rounding to a certain digit and changing
any values less than a certain amount to, e.g., "<.001".
}
