% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownformatrs.R
\name{color}
\alias{color}
\alias{font_col}
\alias{col_code}
\alias{blue}
\alias{dark_blue}
\alias{orange}
\alias{dark_orange}
\alias{dark_gray}
\alias{dark_grey}
\alias{light_gray}
\alias{light_grey}
\alias{white}
\title{Color text in R Markdown}
\usage{
font_col(x, col, code = FALSE)

col_code(x, col)

blue(x, code = FALSE)

dark_blue(x, code = FALSE)

orange(x, code = FALSE)

dark_orange(x, code = FALSE)

dark_gray(x, code = FALSE)

dark_grey(x, code = FALSE)

light_gray(x, code = FALSE)

light_grey(x, code = FALSE)

white(x, code = FALSE)
}
\arguments{
\item{x}{a string, the text to be colored}

\item{col}{the color}

\item{code}{logical. Is this code?}
}
\value{
a character vector
}
\description{
Color HTML or LaTeX text in R Markdown.
}
