% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Search variable labels using regular expressions}
\usage{
dictionary(data, pattern)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{pattern}{the pattern to search for}
}
\value{
a list with locations, names, and labels.
}
\description{
\code{dictionary()} searches variables that have a \code{label} attribute, like those
created with \link[Hmisc:label]{Hmisc::label} or some functions in \link{haven}, for patterns.
Useful for searching large, labelled datasets.
}
