% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatrs.R
\name{est_ci}
\alias{est_ci}
\alias{ci}
\alias{minmax}
\alias{iqr}
\alias{none}
\alias{parenthesis}
\alias{brackets}
\alias{dash}
\alias{comma}
\title{Paste an estimate and CI together}
\usage{
est_ci(est, lower = NULL, upper = NULL, var = NULL,
  include_name = FALSE, bound = parenthesis(), divider = comma(),
  descriptor = ci(95), digits = 2)

ci(lvl = 95)

minmax()

iqr()

none()

parenthesis()

brackets()

dash()

comma()
}
\arguments{
\item{est}{either \code{a data.frame} or a vector of estimates}

\item{lower}{the lower CI. if \code{NULL}, will search for "conf.low"}

\item{upper}{the upper CI. if \code{NULL}, will search for "conf.high"}

\item{var}{estimate name}

\item{include_name}{logical. Should the name of the term be included?}

\item{bound}{bounding symbols around the CI. Default is parentheses.}

\item{divider}{a divider between the upper and lower CI. Default is a comma.}

\item{descriptor}{a descriptor of the CI. Default is "95\% CI".}

\item{digits}{the number of digits to round to}

\item{lvl}{the confidence interval probability}
}
\value{
a \code{data.frame} or vector.
}
\description{
Quickly concatenate an estimate with its confidence intervals. Works with
vectors or output from \code{\link[broom:tidy]{broom::tidy()}}.
}
\examples{

library(broom)
library(dplyr)

x <- tidy(lm(mpg ~ hp, data = mtcars), conf.int = TRUE)
x \%>\%
  mutate(eci = est_ci(estimate, conf.low, conf.high))

est_ci(x)

}
