% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{font_col}
\alias{font_col}
\title{Color text in R Markdown}
\usage{
font_col(x, col)
}
\arguments{
\item{x}{a string, the text to be colored}

\item{col}{the color}
}
\value{
a character vector
}
\description{
Color HTML or LaTeX text in R Markdown.
}
