% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not in}
\usage{
lhs \%not_in\% rhs
}
\arguments{
\item{lhs}{vector or NULL: the values to be matched}

\item{rhs}{vector or NULL: the values to be matched against}
}
\value{
a logical vector
}
\description{
\code{\%not_in\%} is the opposite of \code{\%in\%}. It is a binary operator, which
returns a logical vector indicating if there is not a match of the left hand
side in the right hand side vector.
}
\examples{

1 \%not_in\% 2:10 # TRUE
}
