% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{bound}
\alias{bound}
\alias{sever}
\title{Limit a numeric to be within a certain range.}
\usage{
bound(x, min = NULL, max = NULL)

sever(x, pctl = 99)
}
\arguments{
\item{x}{a numeric vector}

\item{min}{minimum. Ignored if \code{NULL}.}

\item{max}{maximum. Ignored if \code{NULL}.}

\item{pctl}{percentile.}
}
\value{
a numeric vector
}
\description{
\code{bound()} limits a vector to a specific range. \code{sever()} trucates a vector to
a specific quantile.
}
\examples{

x <- rnorm(100, 8, 2)

#  cap at 10
bound(x, max = 10)

# truncate to 95th percentile
sever(x, pctl = 95)

bound(1:100, 25, 75)

}
