% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatrs.R
\name{make_formula}
\alias{make_formula}
\title{Make a formula}
\usage{
make_formula(y, x)
}
\arguments{
\item{y}{the outcome variable}

\item{x}{the exposure variables (a character vector)}
}
\value{
a formula
}
\description{
Quickly make a formula out of character vectors
}
\examples{

make_formula("y", paste0("x", 1:10))
}
