% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_malco}
\alias{palette_malco}
\alias{palette_highlight}
\alias{palette_palitra}
\title{Custom color palettes}
\usage{
palette_malco(index = NULL)

palette_highlight(index = NULL)

palette_palitra(index = NULL)
}
\arguments{
\item{index}{a numeric vector, the colors to include}
}
\value{
a character vector
}
\description{
These palettes are losely based on colorblind-friendly palettes from the
\code{colorblindr} package and colors related to those in the \code{xaringan} package.
}
