% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatrs.R
\name{round_with_zeros}
\alias{round_with_zeros}
\title{Round without dropping zeroes}
\usage{
round_with_zeros(.x, digits = 2)
}
\arguments{
\item{.x}{a numeric vector}

\item{digits}{number of digits to round to}
}
\value{
a character vector
}
\description{
This function rounds a number but forces any dangling zeroes to remain.
}
\examples{

round_with_zeros(1.201)
round_with_zeros(rnorm(10))
}
