% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_malco}
\alias{scale_malco}
\alias{scale_colour_malco}
\alias{scale_color_malco}
\alias{scale_fill_malco}
\alias{scale_edge_colour_malco}
\alias{scale_edge_color_malco}
\alias{scale_edge_fill_malco}
\title{mbmisc ggplot scales}
\usage{
scale_malco(aesthetics, order = 1:7, alpha = NULL, ...)

scale_colour_malco(order = 1:7, alpha = NULL, ...)

scale_color_malco(order = 1:7, alpha = NULL, ...)

scale_fill_malco(order = 1:7, alpha = NULL, ...)

scale_edge_colour_malco(order = 1:7, alpha = NULL, ...)

scale_edge_color_malco(order = 1:7, alpha = NULL, ...)

scale_edge_fill_malco(order = 1:7, alpha = NULL, ...)
}
\arguments{
\item{aesthetics}{What type of aesthetic?}

\item{order}{What's the order of the palette?}

\item{alpha}{A numeric vector}

\item{...}{Passed to \link[ggplot2:discrete_scale]{ggplot2::discrete_scale}}
}
\value{
a ggplot scale
}
\description{
These functions use the colors in \code{palette_malco()} to work easily with
ggplot2 and ggraph.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(mpg, hp, col = factor(am))) +
  geom_point() +
  scale_color_malco()

}
