% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{these_are_real_people_dummy}
\alias{these_are_real_people_dummy}
\title{These are real people, you dummy}
\usage{
these_are_real_people_dummy(.data, x, y, ..., width = 70)
}
\arguments{
\item{.data}{A data frame}

\item{x}{Variable on the x axis}

\item{y}{Variable on the x axis}

\item{...}{Other arguments passed to \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{width}{Width of the subtitle}
}
\value{
a scatter plot
}
\description{
Run this plot every so often to remind yourself that you work with real
people, not data.
}
